Implement IndexedTableSuperclass in rust
This class is vestigial, but serves as a proof of concept. It used to be
a non-generic class to hide a generic/templated IndexedTable, but
IndexedTable is no longer generic.

maturin and pyo3 are used to for rust integration. Install maturin to
use this.

To run:

(Once)
$ python3 -m venv .venv
(Per terminal)
$ source .venv/bin/activate
(On changing the rust code)
$ maturin develop
(Then run python commands as normal)